% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geometry.R
\name{add_geometry}
\alias{add_geometry}
\title{Convert input data frame into a spatial data frame}
\usage{
add_geometry(x, location, feature_type = NA, geom_name = "geometry")
}
\arguments{
\item{x}{Data frame with a feature name column.}

\item{location}{Feature names (tidy evaluation).}

\item{feature_type}{The registered map corresponding to values in \code{location}.
If NA (the default), the type is guessed from the values in \code{location}.}

\item{geom_name}{Name for the new column to contain the geometry.}
}
\value{
A spatial data frame containing all of the columns from the input
data frame.
}
\description{
Convert input data frame into a spatial data frame
}
\examples{
add_geometry(nc_type_example_2, county)
}
