% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cartographer-package.R
\name{cartographer-package}
\alias{cartographer-package}
\title{Interactive Maps with d3-carto-map}
\description{
This package provides an R interface to
\href{https://github.com/emeeks/d3-carto-map}{d3-carto-map} (created by
\href{https://twitter.com/elijah_meeks}{Elijah Meeks}) and the
\href{http://d3js.org/}{d3.js} visualization library (created by Mike
Bostock). It lets you create interactive maps at the R console and in R
documents.
}

