% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cartographer.R
\name{cartographer}
\alias{cartographer}
\title{Create a cartographer map}
\usage{
cartographer(region = NULL, bbox = NULL, width = NULL, height = NULL,
  browser = FALSE)
}
\arguments{
\item{region}{The region to which the map should be centered. This region can
be one of several types. Passing a
\href{http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{two-letter ISO
country code} will center the map on that country. The string \code{"United
States"} will center the map on the continental United States of America.
You can also use the names of continents. The region is case insensitive.}

\item{bbox}{Instead of a region, you can center the map on a bounding box.
The bounding box be specified in decimal degrees of latitude and longitude
and should have the following format: \code{list(c(long1, lat1), c(long2,
lat2))}}

\item{width}{The width of the map in pixels.}

\item{height}{The height of the map in pixels.}

\item{browser}{Where to display the map. If \code{TRUE}, always display in
  the browser. Otherwise, display in the RStudio Viewer if available.}
}
\description{
This function creates a new instance of a d3-carto-map and sets its most
basic options. By itself the function does not create any layers on the map.
}
\examples{
if(require(historydata)) {
  cartographer(region = "United States") \%>\%
    tile_layer() \%>\%
    points_layer(catholic_dioceses, x = "long", y = "lat")
}
}
\seealso{
\code{\link{points_layer}}, \code{\link{tile_layer}},
  \code{\link{topojson_layer}}
}

