% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cartographer.R
\name{cartographerOutput}
\alias{cartographerOutput}
\title{Widget output function for use in Shiny}
\usage{
cartographerOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{ID used for the output.}

\item{width}{Width of the output.}

\item{height}{Height of the output.}
}
\description{
Widget output function for use in Shiny
}

