% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continents_bbox}
\alias{continents_bbox}
\title{Continents bounding boxes}
\format{A data frame with 7 observations of 5 variables.}
\source{
Natural Earth, shapefile for
  \href{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries/}{Admin
   0 - Countries}.
}
\usage{
data(continents_bbox)
}
\description{
A data frame containing bounding coordinates and continent names for seven
continents.
}
\keyword{datasets,}
\keyword{mapping}

