% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographer-global.R
\name{feature_types}
\alias{feature_types}
\title{List known feature types}
\usage{
feature_types()
}
\value{
Character vector of registered feature types.
}
\description{
Each feature type corresponds to map data that has been registered.
}
\examples{
feature_types()
}
\seealso{
\code{\link[=register_map]{register_map()}}
}
