% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographer-global.R
\name{map_outline}
\alias{map_outline}
\title{Retrieve a map outline registered with cartographer.}
\usage{
map_outline(feature_type)
}
\arguments{
\item{feature_type}{Type of map feature. See \code{\link[=feature_types]{feature_types()}} for a list of
registered types.}
}
\value{
The map outline that was registered under \code{feature_type}. Note that
the outline is optional, so this will return \code{NULL} if none was registered.
}
\description{
Retrieve a map outline registered with cartographer.
}
\examples{
map_outline("sf.nc")
}
