% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographer-global.R
\name{map_sf}
\alias{map_sf}
\title{Retrieve map data registered with cartographer.}
\usage{
map_sf(feature_type)
}
\arguments{
\item{feature_type}{Type of map feature. See \code{\link[=feature_types]{feature_types()}} for a list of
registered types.}
}
\value{
The spatial data frame that was registered under \code{feature_type}.
}
\description{
Retrieve map data registered with cartographer.
}
\examples{
map_sf("sf.nc")
}
