% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographer-global.R
\name{map_sfc}
\alias{map_sfc}
\title{Retrieve geometry of a single location.}
\usage{
map_sfc(feature_names, feature_type)
}
\arguments{
\item{feature_names}{Name of the feature(s) to retrieve. This must be an exact
case-sensitive match, and aliases are not consulted.}

\item{feature_type}{Type of map feature. See \code{\link[=feature_types]{feature_types()}} for a list of
registered types.}
}
\value{
The geometry as a \code{sfc} object.
}
\description{
Retrieve geometry of a single location.
}
\examples{
map_sfc("Ashe", "sf.nc")
map_sfc(c("Craven", "Buncombe"), "sf.nc")
}
