% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/points_layer.R
\name{points_layer}
\alias{points_layer}
\title{Points layer}
\usage{
points_layer(carto, data, x = "long", y = "lat", label = "Points",
  color = "red", size = 4, clickable = TRUE, visible = TRUE)
}
\arguments{
\item{visible}{Should the layer be initially visible?}
}
\description{
Add a points layer to a d3-carto-map
}

