% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cartographer.R
\name{renderCartographer}
\alias{renderCartographer}
\title{Widget render function for use in Shiny}
\usage{
renderCartographer(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

