% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sp_to_geojson.R
\name{sp_to_geojson}
\alias{sp_to_geojson}
\title{Convert a spatial data object to a GeoJSON string}
\usage{
sp_to_geojson(sp)
}
\arguments{
\item{sp}{A \code{SpatialPolygonsDataFrame}.}
}
\value{
A string containing unparsed GeoJSON.
}
\description{
Convert an R spatial polygons object to a GeoJSON string that can be used in
\code{\link{geojson_layer}}. The spatial classes are provided by the
\code{sp} package.
}
\note{
This function will write a GeoJSON file to the temporary directory
  created by R, but this file will be deleted when the R session ends.
}
\examples{
if(require(USAboundaries)) {
  # Boundaries of the United States of America in 1800
  us <- us_boundaries(as.Date("1800-01-01"))
  us_geojson <- sp_to_geojson(us)

  cartographer(region = "United States") \%>\%
    geojson_layer(data = us_geojson, clickable = TRUE, label = "US 1800")
}
}

