% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/topojson_layer.R
\name{topojson_layer}
\alias{topojson_layer}
\title{Topojson layer}
\usage{
topojson_layer(carto, data, feature, file = NULL, label = "TopoJSON",
  clickable = FALSE, fill = "lightblue", stroke = "black",
  opacity = 0.5, visible = TRUE)
}
\arguments{
\item{visible}{Should the layer be initially visible?}
}
\description{
Add a topojson layer to a d3-carto-map
}

