% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/world_bounding_boxes.R
\docType{data}
\name{world_bounding_boxes}
\alias{world_bounding_boxes}
\title{World bounding boxes}
\format{A data frame with 236 observations and 6 variables.}
\source{
Natural Earth, shapefile for
  \href{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/10m-admin-0-countries/}{Admin
   0 - Countries}.
}
\usage{
data(world_bounding_boxes)
}
\description{
A data frame containing bounding coordinates and country names and
\href{http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{two-letter ISO codes}
for 236 countries.
}
\keyword{datasets,}
\keyword{mapping}

