% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographer-package.R
\docType{package}
\name{cartographer-package}
\alias{cartographer}
\alias{cartographer-package}
\title{Turn Place Names into Map Data}
\description{
Cartographer is a framework for easily matching spatial data when
you have a list of standardised place names. You might have a data frame
that came from a spreadsheet tracking some data by suburb or state. This
package can convert it into a spatial data frame ready for plotting. The
actual map data is provided by other packages (or your own code) that
register the data with cartographer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cidm-ph/cartographer}
  \item \url{https://cidm-ph.github.io/cartographer/}
  \item Report bugs at \url{https://github.com/cidm-ph/cartographer/issues}
}

}
\author{
\strong{Maintainer}: Carl Suster \email{Carl.Suster@health.nsw.gov.au} (\href{https://orcid.org/0000-0001-7021-9380}{ORCID})

Other contributors:
\itemize{
  \item Western Sydney Local Health District, NSW Health [copyright holder]
}

}
