% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cartographer-package.R, R/cartographer.R
\name{cartographer}
\alias{cartographer}
\title{Interactive Maps with d3-carto-map}
\usage{
cartographer(region = NULL, bbox = list(c(-180, -80), c(180, 80)),
  width = NULL, height = NULL)
}
\description{
Interactive Maps with d3-carto-map

<Add Description>
}

