% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_type_example_1}
\alias{nc_type_example_1}
\alias{nc_type_example_2}
\title{Example datasets with a feature name column and random data}
\format{
A data frames with 50 rows and 2 columns:
\describe{
\item{county}{Feature names that match the NAME field of the nc dataset}
\item{type}{Arbitrary categorical data}
}

A data frames with 200 rows and 2 columns:
}
\usage{
nc_type_example_1

nc_type_example_2
}
\description{
This dataset contains random data compatible with the \code{sf.nc} example
map data for illustrating cartographer's features. \code{nc_type_example_1}
contains a deliberate error in the county name for a single row, whereas
\code{nc_type_example_2} contains correct data.
}
\keyword{datasets}
