% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/points_layer.R
\name{points_layer}
\alias{points_layer}
\title{Points layer}
\usage{
points_layer(carto, data, x = "long", y = "lat", label = "Points",
  color = "red", size = 4, radius_field = NULL, clickable = TRUE,
  visible = TRUE, cluster = FALSE,
  radius_function = "d3.scale.sqrt().domain([0, layer.radius_domain]).range([0, layer.size])")
}
\arguments{
\item{carto}{The map created by \code{cartographer()}.}

\item{data}{A data frame with columns for the latitude and longitude in
decimal degrees. It may also contain other columns to be used as
properties.}

\item{x}{The name of the column in the data frame with the longitudes.}

\item{y}{The name of the column in the data frame with the latitudes.}

\item{label}{The label for this layer on the layer selector.}

\item{color}{The color of the points, either as a name recognized by a
browser or as a hexadecimal string (e.g., \code{#4527f3}).}

\item{size}{The radius of the points. If the parameter \code{radius_field} is
set so that the points are scaled to a variable in \code{data}, then the
\code{size} parameter controls the maximum size of the points.}

\item{radius_field}{The name of a numeric field in the data frame passed to
\code{data}. This field will be used to scale the points. The domain of the
scale will run from 0 to the maximum value in the \code{radius_field}; the
range will run from 0 to the value in \code{size}.}

\item{clickable}{Controls whether a point can be clicked to show the
properties in the data frame.}

\item{visible}{Should the layer be initially visible?}

\item{cluster}{If true, when the map is zoomed points will be clustered on a
quadtree. If the points are clustered, arguments that control the size of
points will be disregarded.}

\item{radius_function}{A
\href{https://github.com/mbostock/d3/wiki/Quantitative-Scales}{D3
quantitative scale function} to control the radius of the points.}
}
\description{
Add a points layer to a d3-carto-map.
}
\examples{
if(require(historydata)) {
  cartographer(region = "United States") \%>\%
    tile_layer() \%>\%
    points_layer(data = paulist_missions, x = "long", y = "lat",
                 color = "blue", label = "Missions",
                 radius_field = "confessions", size = 8)
}
}

